requirements.txt
altair==5.4.1
asttokens==2.4.1
attrs==23.2.0
blinker==1.8.2
cachetools==5.5.0
category==0.0.3
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
colorlover==0.3.0
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
debugpy==1.8.7
decorator==5.1.1
encoders==0.0.4
executing==2.1.0
fastjsonschema==2.19.1
fonttools==4.51.0
gitdb==4.0.11
GitPython==3.1.43
idna==3.10
imbalanced-learn==0.12.4
imblearn==0.0
ipykernel==6.29.5
ipython==8.28.0
ipython-genutils==0.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets==3.0.10
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
narwhals==1.10.0
nest-asyncio==1.6.0
numpy==2.1.2
packaging==24.1
pandas==2.2.3
parso==0.8.4
patsy==0.5.6
pillow==10.3.0
platformdirs==4.3.6
plotly==5.22.0
prompt_toolkit==3.0.48
protobuf==5.28.3
psutil==6.0.0
pure_eval==0.2.3
pyarrow==17.0.0
pydeck==0.9.1
Pygments==2.18.0
pyngrok==7.2.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.2
pywin32==308
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
regex==2024.4.16
requests==2.32.3
rich==13.9.3
rpds-py==0.18.1
scikit-learn==1.4.2
scipy==1.13.0
seaborn==0.13.2
setuptools==70.0.0
six==1.16.0
smmap==5.0.1
stack-data==0.6.3
streamlit==1.39.0
tenacity==8.3.0
threadpoolctl==3.4.0
toml==0.10.2
tornado==6.4.1
tqdm==4.66.2
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
watchdog==5.0.3
wcwidth==0.2.13
widgetsnbextension==4.0.10
xgboost==2.1.2
